+++ nedsettings_form.php
@@ -85,6 +85,24 @@ class course_ned_edit_form extends moodleform {
             unset($choices);
         }
 
+        $viewjumptomenuoptions = array(
+            0 => get_string('everyone', 'format_ned'),
+            1 => get_string('courseeditors', 'format_ned')
+        );
+        $mform->addElement('select', 'viewjumptomenu',
+            get_string('viewjumptomenu', 'format_ned'), $viewjumptomenuoptions
+        );
+        unset($viewjumptomenuoptions);
+
+        $viewsectionforwardbacklinksoptions = array(
+            0 => get_string('everyone', 'format_ned'),
+            1 => get_string('courseeditors', 'format_ned')
+        );
+        $mform->addElement('select', 'viewsectionforwardbacklinks',
+            get_string('viewsectionforwardbacklinks', 'format_ned'), $viewsectionforwardbacklinksoptions
+        );
+        unset($viewsectionforwardbacklinksoptions);
+
         $this->add_action_buttons();
     }